+++ R/models.R
@@ -1053,8 +1053,8 @@ stlm_arima <- function(train_data,
 
   model_spec_stlm_arima <- modeltime::seasonal_reg(
     seasonal_period_1 = seasonal_period_stlm_arima[1],
-    seasonal_period_2 = seasonal_period_stlm_arima[2],
-    seasonal_period_3 = seasonal_period_stlm_arima[3]
+    seasonal_period_2 = if(is.na(seasonal_period_stlm_arima[2])) {NULL} else {seasonal_period_stlm_arima[2]},
+    seasonal_period_3 = if(is.na(seasonal_period_stlm_arima[3])) {NULL} else {seasonal_period_stlm_arima[3]}
   ) %>%
     parsnip::set_engine("stlm_arima")
 
@@ -1082,8 +1082,8 @@ stlm_ets <- function(train_data,
 
   model_spec_stlm_ets <- modeltime::seasonal_reg(
     seasonal_period_1 = seasonal_period_stlm_ets[1],
-    seasonal_period_2 = seasonal_period_stlm_ets[2],
-    seasonal_period_3 = seasonal_period_stlm_ets[3]
+    seasonal_period_2 = if(is.na(seasonal_period_stlm_ets[2])) {NULL} else {seasonal_period_stlm_ets[2]},
+    seasonal_period_3 = if(is.na(seasonal_period_stlm_ets[3])) {NULL} else {seasonal_period_stlm_ets[3]}
   ) %>%
     parsnip::set_engine("stlm_ets")
 
@@ -1273,8 +1273,8 @@ tbats <- function(train_data,
 
   model_spec_tbats <- modeltime::seasonal_reg(
     seasonal_period_1 = seasonal_period_tbats[1],
-    seasonal_period_2 = seasonal_period_tbats[2],
-    seasonal_period_3 = seasonal_period_tbats[3]
+    seasonal_period_2 = if(is.na(seasonal_period_tbats[2])) {NULL} else {seasonal_period_tbats[2]},
+    seasonal_period_3 = if(is.na(seasonal_period_tbats[3])) {NULL} else {seasonal_period_tbats[3]}
   ) %>%
     parsnip::set_engine("tbats")